fix: split Microsoft 365/Hotmail functionality
`hotmail` was previously behind the `headless` feature which, if
disabled, breaks the Microsoft 365 API functionality.

refactor these into a `microsoft` module, leaving the `hotmail` module
behind the `headless` feature and allowing the Microsoft 365 API to
exist separately.
