+++ core/src/smtp/microsoft/hotmail.rs
@@ -23,15 +23,10 @@ use fantoccini::{
 	ClientBuilder, Locator,
 };
 use futures::TryFutureExt;
-use reqwest::Error as ReqwestError;
 use serde::Serialize;
 use serde_json::Map;
 
-use super::SmtpDetails;
-use crate::LOG_TARGET;
-use crate::{
-	smtp::http_api::create_client, util::ser_with_display::ser_with_display, CheckEmailInput,
-};
+use crate::{smtp::SmtpDetails, util::ser_with_display::ser_with_display, LOG_TARGET};
 
 #[derive(Debug, Serialize)]
 pub enum HotmailError {
@@ -39,8 +34,6 @@ pub enum HotmailError {
 	Cmd(CmdError),
 	#[serde(serialize_with = "ser_with_display")]
 	NewSession(NewSessionError),
-	#[serde(serialize_with = "ser_with_display")]
-	ReqwestError(ReqwestError),
 }
 
 impl From<CmdError> for HotmailError {
@@ -52,12 +45,6 @@ impl From<CmdError> for HotmailError {
 impl From<NewSessionError> for HotmailError {
 	fn from(e: NewSessionError) -> Self {
 		Self::NewSession(e)
-	}
-}
-
-impl From<ReqwestError> for HotmailError {
-	fn from(error: ReqwestError) -> Self {
-		HotmailError::ReqwestError(error)
 	}
 }
 
@@ -151,65 +138,9 @@ pub async fn check_password_recovery(
 	})
 }
 
-/// Convert an email address to its corresponding OneDrive URL.
-fn get_onedrive_url(email_address: &str) -> String {
-	let (username, domain) = email_address
-		.split_once('@')
-		.expect("Email address syntax already validated.");
-	let (tenant, _) = domain
-		.split_once('.')
-		.expect("Email domain syntax already validated.");
-
-	format!(
-		"https://{}-my.sharepoint.com/personal/{}_{}/_layouts/15/onedrive.aspx",
-		tenant,
-		username.replace('.', "_"),
-		domain.replace('.', "_"),
-	)
-}
-
-/// Use a HTTP request to verify if an Microsoft 365 email address exists.
-///
-/// See
-/// [this article](<https://www.trustedsec.com/blog/achieving-passive-user-enumeration-with-onedrive/>)
-/// for details on the underlying principles.
-///
-/// Note that a positive response from this function is (at present) considered
-/// a reliable indicator that an email-address is valid. However, a negative
-/// response is ambigious: the email address may or may not be valid but this
-/// cannot be determined by the method outlined here.
-pub async fn check_microsoft365_api(
-	to_email: &EmailAddress,
-	input: &CheckEmailInput,
-) -> Result<Option<SmtpDetails>, HotmailError> {
-	let url = get_onedrive_url(to_email.as_ref());
-
-	let response = create_client(input, "microsoft365")?
-		.head(url)
-		.send()
-		.await?;
-
-	log::debug!(
-		target: LOG_TARGET,
-		"[email={}] microsoft365 response: {:?}",
-		to_email,
-		response
-	);
-
-	if response.status() == 403 {
-		Ok(Some(SmtpDetails {
-			can_connect_smtp: true,
-			is_deliverable: true,
-			..Default::default()
-		}))
-	} else {
-		Ok(None)
-	}
-}
-
 #[cfg(test)]
 mod tests {
-	use super::{check_password_recovery, get_onedrive_url};
+	use super::check_password_recovery;
 	use async_smtp::EmailAddress;
 	use async_std::prelude::FutureExt;
 	use std::str::FromStr;
@@ -259,13 +190,5 @@ mod tests {
 
 		let f = f1.try_join(f2).await;
 		assert!(f.is_ok(), "{:?}", f);
-	}
-
-	#[test]
-	fn test_onedrive_url() {
-		let email_address = "lightmand@acmecomputercompany.com";
-		let expected = "https://acmecomputercompany-my.sharepoint.com/personal/lightmand_acmecomputercompany_com/_layouts/15/onedrive.aspx";
-
-		assert_eq!(expected, get_onedrive_url(email_address));
 	}
 }
