+++ core/src/smtp/microsoft/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "headless")]
+pub mod hotmail;
+pub mod microsoft365;
